Create data folder and move all files in it
