+++ lib/data/repositories/auth_repository_impl.dart
@@ -1,15 +1,18 @@
-import 'package:aspdm_project/model/user.dart';
+import 'package:aspdm_project/domain/entities/user.dart';
+import 'package:aspdm_project/domain/repositories/auth_repository.dart';
 import 'package:aspdm_project/services/data_source.dart';
 import 'package:aspdm_project/services/preference_service.dart';
 
-import '../locator.dart';
+import '../../locator.dart';
 
-class AuthRepository {
+class AuthRepositoryImpl extends AuthRepository {
   DataSource _dataSource = locator<DataSource>();
   PreferenceService _preferenceService = locator<PreferenceService>();
 
+  @override
   User get lastSignedInUser => _preferenceService.getLastSignedInUser();
 
+  @override
   Future<User> login(String email, String password) async {
     final user = await _dataSource.authenticate(email, password);
     if (user == null) throw Exception("Invalid user!");
@@ -18,6 +21,7 @@ class AuthRepository {
     return user;
   }
 
+  @override
   Future<void> logout() {
     return _preferenceService.storeSignedInUser(null);
   }
